Move utility methods to a module to reduce load_into method complexity
