+++ spec/spec_helper.rb
@@ -1,6 +1,5 @@
 require 'capistrano-spec'
 require 'capistrano-unicorn'
-require 'pry'
 
 RSpec.configure do |config|
   config.include Capistrano::Spec::Matchers
